backend/app/requirements.txt
aiohttp==3.12.9
asyncio==3.4.3
docling==2.31.0
docling-core==2.30.0
fastapi==0.115.12
fastapi-cli==0.0.7
httpx==0.28.1
langfuse==3.0.0
llama-index==0.12.41
llama-index-embeddings-huggingface==0.5.3
llama-index-llms-openrouter==0.3.2
llama-index-vector-stores-postgres==0.5.2
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
openinference-instrumentation==0.1.32
openinference-instrumentation-llama-index==4.3.0
partial-json-parser==0.2.1.1.post5
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg-pool==3.2.6
pyarrow==20.0.0
pydantic==2.11.5
sse-starlette==2.3.4
itsdangerous==2.2.0
authlib==1.6.0
jwt==1.3.1